refactor(install): move RecipeRecommender interface closer to its usage
